test: e2e: topomgr: move test-specific utilites
This patch prepares for the further generalization of
`numa_alignment.go` moving test-specific utilities, which
aren't yet useful to other tests, closer to their clients.

This way we enable a future patch to move `numa_alignment.go`
into a proper utility module.

No intended changes in behaviour.
